@user202729 2024-10-28
Maybe add a few tests?

```
sage: matrix(GF(2), 2, 0).solve_right(matrix(GF(2), 2, 2))
[]
sage: matrix(GF(2), 2, 0).solve_right(matrix(GF(2), 2, 2)+1)
Traceback…

ValueError: matrix equation has no solutions
sage: matrix(GF(2), 2, 0).solve_right(matrix(GF(2), 2, 2)+1, check=False)
[]
sage: matrix(GF(2), 2, 2).solve_right(matrix(GF(2), 2, 0))
[]
```

(note that the old code returns empty matrix for the third case, while the new code raises error. Apparently it doesn't break any existing test case, but anyway.)