@maple3142 2024-11-04
Tests for special case added.

> Strange that m4ri doesn't automatically handle the empty matrix case correctly (mathematically) though. Any idea why?

TBH, zero size matrix is pretty weird by itself I think, so it is not unusual to not consider that I think. It would probably be better handled by the `solve_right` in `matrix2.py` so the underlying function don't need to handle that.