@maple3142 2024-11-05
Not sure what breaks the test using Meson on Python 3.9...