@user202729 2024-11-05
My argument: just as zero is not a weird integer, zero-dimensional vector space (which is just a point) is not a weird space. So a zero-sized matrix is just a linear map from (or to) a zero-dimensional vector space.

The math checks out.

(Of course, programming-wise it is still a corner case.