@user202729 2024-11-08
Looks like the failing test is https://github.com/sagemath/sage/pull/38940 . Should be unrelated to this change then.