@grhkm21 2024-11-25
Thanks
love `# although it is called mzd_solve_left, it does the same thing as solve_right`